requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
appnope==0.1.4
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
autopep8==2.3.2
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
blis==1.3.0
build==1.2.2.post1
catalogue==2.0.10
cbor2==5.6.5
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.1
confection==0.1.5
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==45.0.4
cymem==2.0.11
daphne==4.2.0
decorator==5.2.1
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==3.0.0
Django==5.2.3
django-celery-beat==2.8.1
django-channels==0.7.0
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-countries==7.6.1
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-import-export==4.3.7
django-ipware==7.0.1
django-js-asset==3.1.2
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-picklefield==3.3
django-ratelimit==4.1.0
django-redis==5.4.0
django-rq==3.0.1
django-taggit==6.1.0
django-timezone-field==7.1
django-tinymce==4.1.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework-queryfields==1.1.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docopt==0.6.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
executing==2.2.0
factory_boy==3.3.3
Faker==37.4.0
fastjsonschema==2.21.1
fido2==2.0.0
flake8==7.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
geoip2==5.1.0
greenlet==3.2.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipython>=8.12,<10
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
maxminddb==2.7.0
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.0
multidict==6.4.4
murmurhash==1.0.13
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pendulum==3.1.0
pexpect==4.9.0
phonenumbers==9.0.7
pickleshare==0.7.5
pillow==11.2.1
pip-review==1.3.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.3.8
pluggy==1.6.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.13.1
pynvim==0.5.2
pyOpenSSL==25.1.0
pyotp==2.9.0
pyproject_hooks==1.2.0
pytest==8.4.0
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==8.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rq==2.3.3
sentry-sdk==2.30.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
tablib==3.8.0
thinc==8.3.6
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
twilio==9.6.3
Twisted==25.5.0
txaio==23.1.1
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webauthn==2.5.2
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarg==0.1.9
yarl==1.20.1
zope.interface==7.2
